Moving links for all Jekyll-generated pages to `_includes/links.md`.
1.  Create file `_includes/links.md` with definitions of all links.
2.  Replace links at the bottom of Jekyll'd `.md` files with inclusion of that file.
3.  Modify `bin/lesson_check.py` to load and check definitions in this file.

This eliminates most redundancy in link definitions, though
`CONTRIBUTING.md` and `README.md` still contain their own links, since
they are not Jekyll'd.  Note that we *cannot* put the links into
`_layouts/*.html` (or even `_layouts/something.md`) because of the
order in which Jekyll expands things - we have to explicitly include
it in each page.